+++ src/components/app-icon.tsx
@@ -2,11 +2,11 @@ import { platform } from "@tauri-apps/plugin-os";
 import { AppWindow, AppWindowMac, LucideProps } from "lucide-react";
 import { forwardRef, ForwardRefExoticComponent } from "react";
 
-export const WindowIcon: ForwardRefExoticComponent<Omit<LucideProps, "ref"> & React.RefAttributes<SVGSVGElement>> = forwardRef(props=>{
+export const WindowIcon: ForwardRefExoticComponent<Omit<LucideProps, "ref"> & React.RefAttributes<SVGSVGElement>> = forwardRef((props,ref)=>{
      const currentPlatform = platform();
      return currentPlatform==="macos" ? (
-          <AppWindowMac {...props}/>
+          <AppWindowMac {...props} ref={ref}/>
      ) : (
-          <AppWindow {...props}/>
+          <AppWindow {...props} ref={ref}/>
      )
 })